Add an Appmattus (with OkHttp) CT-checked request
This seems to be the most common case of Appmattus, which is actually
quite widely used. It's not really certificate pinning, but certificate
transparency has better real world use cases, and causes equal problems
for MitM interception, so it's well worth including here as a test case.

Appmattus can be used in multiple scenarios, but OkHttp seems most
recommended and popular, and should be representative of most cases
anyway.