Directly trust the LE R3 pin too, to fix manual case
Seems that badssl.com has had some changes and no longer includes ISRG.
The OkHTTP case seems to check the full chain (presumably from local
stores) but that's complicated to do ourselves. This seems like a
reasonable short-term fix, and it's easy to patch up later on now that
we know what's going on.